refactor: 💡 Chart tooltips refactor
