--- packages/charts/src/charts/bar/components/bar-tooltip-content/index.ts
@@ -1,3 +1,0 @@
-import BarTooltipContent from './bar-tooltip-content.component';
-
-export default BarTooltipContent;
