+++ packages/charts/src/charts/bar/components/index.ts
@@ -1,6 +1,5 @@
 import Bar from './bar';
-import BarTooltipContent from './bar-tooltip-content';
 import BarValues from './bar-values';
 import Bars from './bars';
 
-export { Bar, BarTooltipContent, BarValues, Bars };
+export { Bar, BarValues, Bars };
