+++ packages/charts/src/charts/heatmap/components/block/block.styles.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
 
-import { SHADOW_FILTER_ID } from './constants';
+import { SHADOW_FILTER_ID } from '../../constants';
 
 export const StyledRect = styled.rect<{
   dropShadow: boolean;
