+++ packages/charts/src/charts/heatmap/components/heatmap/heatmap.component.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Layout } from '@keen.io/ui-core';
 import { AnimatePresence } from 'framer-motion';
 
-import Block from './block.component';
+import Block from '../block/block.component';
 
-import { BlockType } from './types';
+import { BlockType } from '../../types';
 
 type Props = {
   blocks: BlockType[];
