--- packages/charts/src/charts/line/components/tooltip/index.ts
@@ -1,3 +1,0 @@
-import Tooltip from './tooltip.component';
-
-export default Tooltip;
