+++ packages/charts/src/components/distributed-chart-tooltip/index.ts
@@ -1,0 +1,8 @@
+import DistributedChartTooltip from './distributed-chart-tooltip.component';
+import {
+  getLabel,
+  getSelectors,
+} from './utils/distributed-chart-tooltip-utils';
+
+export { getLabel, getSelectors };
+export default DistributedChartTooltip;
